mm/cma: always check which page caused allocation failure
Now, we have tracepoint in test_pages_isolated() to notify which pfn
cannot be isolated.  But in alloc_contig_range(), some error paths don't
call test_pages_isolated() so it's still hard to know the exact pfn that
caused allocation failure.

This patch changes this situation by calling test_pages_isolated() in
error paths.  In the allocation failure case, some overhead is added by
this change, but allocation failure is a really rare event so it does not
matter.

In the fatal signal pending case, we don't call test_pages_isolated()
because this failure is an intentional one.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Michal Nazarewicz <mina86@mina86.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>