ocfs2/dlm: fix a race between purge and migration
We found a race between purge and migration when doing code review.  Node
A put lockres to purgelist before receiving the migrate message from node
B which is the master.  dlm_mig_lockres_handler finds the lockres and
releases the dlm spinlock.  Then dlm_thread on node A can purge the
lockres.  dlm_mig_lockres_handler then gets the lockres spinlock and sends
assert master message to tell other nodes that node A is the master now
even lockres is purged.  If node C set the master to node A, and another
node D master the lockres because no node respond he is the master.  That
will make node C crash when node D send assert master to node C.  So check
if lockres gets unhashed in dlm_mig_lockres_handler to fix this race.

Signed-off-by: Jiufei Xue <xuejiufei@huawei.com>
Reviewed-by: Joseph Qi <joseph.qi@huawei.com>
Reviewed-by: Yiwen Jiang <jiangyiwen@huawei.com>
Cc: Mark Fasheh <mfasheh@suse.de>
Cc: Joel Becker <jlbec@evilplan.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>