mm, dax, pmem: introduce {get|put}_dev_pagemap() for dax-gup
get_dev_page() enables paths like get_user_pages() to pin a dynamically
mapped pfn-range (devm_memremap_pages()) while the resulting struct page
objects are in use.  Unlike get_page() it may fail if the device is, or is
in the process of being, disabled.  While the initial lookup of the range
may be an expensive list walk, the result is cached to speed up subsequent
lookups which are likely to be in the same mapped range.

devm_memremap_pages() now requires a reference counter to be specified at
init time.  For pmem this means moving request_queue allocation into
pmem_alloc() so the existing queue usage counter can track "device pages".

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Tested-by: Logan Gunthorpe <logang@deltatee.com>
Cc: Dave Hansen <dave.hansen@linux.intel.com>
Cc: Matthew Wilcox <willy@linux.intel.com>
Cc: Ross Zwisler <ross.zwisler@linux.intel.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>