mm/memblock: introduce for_each_memblock_type()
We already have the for_each_memblock() macro in <linux/memblock.h> which
provides ability to iterate over memblock regions of a known type.  The
for_each_memblock() macro allows us to pass the pointer to the struct
memblock_type, instead we need to pass name of the type.

This patch introduces a new macro for_each_memblock_type() which allows us
iterate over memblock regions with the given type when the type is
unknown.

Signed-off-by: Alexander Kuleshov <kuleshovmail@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>