dynamic_debug: add jump label support
Although dynamic debug is often only used for debug builds, sometimes its
enabled for production builds as well. Minimize its impact by using jump
labels. This reduces the text section by 6000+ bytes in the kernel image
below. It does increae data, but this should only be referenced when
changing the direction of the branches, and hence usually not in cache.

    text     data     bss      dec    hex filename
10369421  2033336  917504 13320261 cb4045 vmlinux.pre
10363054  2103736  917504 13384294 cc3a66 vmlinux.post

Signed-off-by: Jason Baron <jbaron@akamai.com>
Cc: Joe Perches <joe@perches.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>